requirements.txt
SpeechRecognition
gtts
playsound
request
colorama